requirements.txt
fastapi
uvicorn
streamlit
pandas
psycopg[binary]
pgvector
openai
python-dotenv
pydantic
sqlalchemy
